fix: comprehensive boilerplate review fixes across scaffold, code quality, and testing
Resolves 27 issues (9 critical, 12 important, 6 minor) across shared infrastructure,
domain/app layer, adapters, scaffold, and documentation:

Shared Infrastructure:
- fix(errors): add Key field to DomainError for identity matching
- fix(connectutil): preserve error chain in DomainErrorToConnect
- fix(middleware): atomic Lua rate limiter implementation
- fix(middleware): remove cross-module RBAC import
- fix(events): make DLQ declaration failure fatal on startup

User Module:
- fix(user/domain): improve error constructors and event documentation
- fix(user/app): reset changedFields on closure retry, fix delete fallback
- fix(user/adapters/postgres): single window-function query for ListUsers, clamp page size
- fix(user/adapters/grpc): fix response mapping and handler contracts
- fix(user/adapters/search): improve noop test organization

Support Services:
- fix(notification/smtp): add 30s hard timeout to prevent goroutine leak
- fix(scaffold): fix rollback RemoveAll, add TotalPages to templates

Documentation:
- docs(architecture): clarify hexagonal patterns and layer responsibilities
- docs(code-standards): fix naming conventions and error handling guidance
- docs(adding-a-module): update scaffold command and module structure references

+++ internal/modules/user/domain/repository.go
@@ -12,11 +12,11 @@ type ListResult struct {
 }
 
 // TotalPages computes the total number of pages for the given page size.
-func (r ListResult) TotalPages(pageSize int) int64 {
+func (r ListResult) TotalPages(pageSize int) int {
 	if pageSize <= 0 {
 		return 0
 	}
-	return (int64(r.Total) + int64(pageSize) - 1) / int64(pageSize)
+	return (r.Total + pageSize - 1) / pageSize
 }
 
 // UserRepository is the port for user persistence.
